Compiles with gcc 4.4.3

+++ BASim/src/Core/Timer.hh
@@ -87,7 +87,7 @@ public:
   /// returns the number of timings elapsed time represents
   int getCumulativeCount()
   {
-    return cumulativeCount;
+    return static_cast<int>(cumulativeCount);
   }
 
   int getNestingLevel()
